Make variables funciton scoped to match python.
This simplifies a lot of code. Especially when mixing atomic blocks within serial code. Previously, variables declared within atomic blocks are not visible outside. So, we need to do a lot of workarounds. This would simplify that.

For the state variables, we still use only the top level variables within the init functiosn as state variables instead of all variables in the init

Co-authored-by: jayaprabhakar <jayaprabhakar@gmail.com>